+++ CHANGELOG.md
@@ -17,7 +17,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * `iBlock`:
   * Created separate buckets for each `loadModules` call
-  * Fixed wait in `loadModules`: now modules are not loaded until wait is resolved
+  * Fixed wait in `loadModules`: now, modules are not loaded until the wait is resolved
 
 ## v4.0.0-beta.64 (2024-02-19)