+++ src/components/super/i-block/CHANGELOG.md
@@ -14,7 +14,7 @@ Changelog
 #### :bug: Bug Fix
 
 * Created separate buckets for each `loadModules` call
-* Fixed wait in `loadModules`: now modules are not loaded until wait is resolved
+* Fixed wait in `loadModules`: now, modules are not loaded until the wait is resolved
 
 ## v4.0.0-beta.64 (2024-02-19)